Do not always destroy the box

+++ scripts/build_ci.sh
@@ -14,8 +14,12 @@ DIR=`dirname $0`
 
 cd $DIR/..
 
+
+
 # Re-create vagrant box
-vagrant destroy --force
+if  [ "$DESTROY_BOX" = true ] ; then
+    vagrant destroy --force
+fi
 
 # make sure Gitlab is known to the runner
 GITLAB_HOST='gitlab.liip.ch'
@@ -29,7 +33,14 @@ git submodule update --recursive
 vagrant plugin install vagrant-omnibus
 vagrant plugin install vagrant-triggers
 
-vagrant up
+# check if there is already a vagrant box
+vtext=`vagrant status 2>/dev/null | awk '{$1=""; print $0}' | sed 's/^ //g' | grep virtualbox` 
+if [[ $vtext =~ .*(running|poweroff).* ]]
+then
+    vagrant reload --provision
+else
+    vagrant up
+fi
 
 # Run build script in the vagrant box
 vagrant ssh -c /vagrant/scripts/build.sh
